# Conflicts:
#	OsmAnd-java/src/main/java/net/osmand/router/network/NetworkRouteContext.java
#	OsmAnd-java/src/main/java/net/osmand/router/network/NetworkRouteSelector.java
#	OsmAnd/src/net/osmand/plus/onlinerouting/EngineParameter.java
#	OsmAnd/src/net/osmand/plus/onlinerouting/engine/GpxEngine.java
#	OsmAnd/src/net/osmand/plus/onlinerouting/engine/OnlineRoutingEngine.java
#	OsmAnd/src/net/osmand/plus/onlinerouting/ui/OnlineRoutingEngineFragment.java
#	OsmAnd/src/net/osmand/plus/profiles/SelectOnlineApproxProfileBottomSheet.java
#	OsmAnd/src/net/osmand/plus/track/helpers/NetworkRouteSelectionTask.java

+++ OsmAnd/src/net/osmand/plus/mapcontextmenu/MapContextMenuFragment.java
@@ -1255,7 +1255,7 @@ public class MapContextMenuFragment extends BaseOsmAndFragment implements Downlo
 				TextView progressTitle = view.findViewById(R.id.progressTitle);
 				progressTitle.setText(titleProgressController.caption);
 				progressBar.setIndeterminate(titleProgressController.indeterminate);
-				progressBar.setProgress(titleProgressController.progress);
+				progressBar.setProgress((int) titleProgressController.progress);
 				progressBar.setVisibility(titleProgressController.progressVisible ? View.VISIBLE : View.GONE);
 
 				ImageView progressButton = view.findViewById(R.id.progressButton);
